Final Files
In this project we are going to develop a game. The Zombies apocalypse is a 3D game, In which the user can Play in the first-person perspective. Idea of the game is the school building staff and students have turned into zombies. These zombies eat live humans. The character is a live student of a school, and the player needs to navigate through levels to defend himself in the hope of getting out of the building alive.